container stats

+++ main/src/main/java/net/oneandone/stool/docker/Stats.java
@@ -3,10 +3,10 @@ package net.oneandone.stool.docker;
 public class Stats {
     /* cpu percentage */
     public final int cpu;
-    public final int memoryUsage;
-    public final int memoryLimit;
+    public final long memoryUsage;
+    public final long memoryLimit;
 
-    public Stats(int cpu, int memoryUsage, int memoryLimit) {
+    public Stats(int cpu, long memoryUsage, long memoryLimit) {
         this.cpu = cpu;
         this.memoryUsage = memoryUsage;
         this.memoryLimit = memoryLimit;
